mais ajustes

+++ motorapido/src/br/com/motorapido/dao/IMensagemMotoristaFuncionarioDAO.java
@@ -4,7 +4,7 @@ import br.com.minhaLib.dao.CriterioOrdenacao;
 import br.com.minhaLib.dao.GenericDAO;
 import br.com.motorapido.entity.MensagemMotoristaFuncionario;
 
-public interface IMensagemMotoristaFuncionarioDAO extends GenericDAO<MensagemMotoristaFuncionario, Integer> {
+public interface IMensagemMotoristaFuncionarioDAO extends GenericDAO<MensagemMotoristaFuncionario, Long> {
 
 	public static CriterioOrdenacao BY_DTCRIACAO_ASC = CriterioOrdenacao
 			.asc("dataCriacao");
